successfully map card to fragment. other cards are commented out for now.
