+++ NowFeed/app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
         <activity
             android:name=".MapsActivity"
             android:label="@string/app_name"
-            android:screenOrientation="portrait">
+            >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
