+++ NowFeed/app/src/main/java/com/jaellysbales/nowfeed/DetailsFragment.java
@@ -1,0 +1,9 @@
+package com.jaellysbales.nowfeed;
+
+import android.app.Fragment;
+
+/**
+ * Created by charlynbuchanan on 7/25/15.
+ */
+public class DetailsFragment extends Fragment {
+}
