+++ NowFeed/app/src/main/res/layout/weather_card.xml
@@ -2,6 +2,7 @@
 <android.support.v7.widget.CardView
     xmlns:card_view="http://schemas.android.com/apk/res-auto"
     xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="250dp"
     card_view:cardBackgroundColor="@color/darkTeal"
@@ -98,7 +99,8 @@
             android:layout_alignParentBottom="true"
             android:layout_alignParentRight="true"
             android:layout_alignParentEnd="true"
-            android:textColor="@color/cream"/>
+            android:textColor="@color/cream"
+            tools:ignore="RtlCompat" />
 
         <TextView
             android:layout_width="wrap_content"
